add detection and tracking

+++ detect_in_camera/test/test.cpp
@@ -143,8 +143,10 @@ int main(int argc, char **argv) {
 				right_detected = false;
 			}
 			if (!tracker_initialized && left_detected && right_detected) {//initializes the tracker//only do one time
-				target_left_box = (target_left[0].x,target_left[0].y, target_left[0].width, target_left[0].height);
-				target_right_box = (target_right[0].x,target_right[0].y, target_right[0].width, target_right[0].height);
+				target_left_box = ((double)target_left[0].x,(double)target_left[0].y,
+						(double)target_left[0].width, (double)target_left[0].height);
+				target_right_box = ((double)target_right[0].x,(double)target_right[0].y,
+						(double)target_right[0].width, (double)target_right[0].height);
 				if (!tracker_left->init(pic_left, target_left_box)) {
 					cout << "***Could not initialize left tracker...***\n";
 					left_detected = false;
